src/backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
click==8.3.1
fastapi==0.104.1
greenlet==3.3.1
h11==0.16.0
httptools==0.7.1
idna==3.11
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dotenv==1.0.0
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
